requirements.txt
flask
flask-sqlalchemy
flask-login
flask-mail
werkzeug
google-cloud-aiplatform
stripe
requests
python-dotenv